Handle bad newsgroups header
This requires the matching update to STUMP to add a warning header.

Add a management page to allow each group administrator to control what
happens when the incoming message had a bad Newsgroups header. The three
cases of header missing, empty or not naming the target group can have
their actions controlled separately. Options are to just fix the header
as was done before this change, add a warning, or reject for any of the
available rejection reasons.

Incoming messages with the warning header added by STUMP will be
processed as specified by the group administrator.