Use textarea for the moderator comment
A textarea for the moderator comment means that the moderator can enter
a long comment more easily. It also means that the moderator can enter
several lines because RETURN does not submit the form.

The textarea will wrap lines for display but they are sent as typed so
this should be used in conjunction with the fix for STUMP that wraps
long lines in the incoming comment.