Rework incoming message handling for MIME
This replaces the MIME handling code for incoming messages. When the
use_mime option is set to 'yes' this uses MIME::Parser to do more of the
work. More information is captured about the parts and this is used to
drive the display of the article. The replacement code has been put into
a new module with the script directory location being added to @INC so
that it can be found in the conventional way.

The use of global variable has been reduced in some places and explicit
declarations added elsewhere.

The article display code is now shared by the multiple article and
single article forms. The data part of the prolog file is no longer
needed