Buttons to refresh or go back to main forms
This replaces the link at the bottom of the page that took you back to
the welcome page with one or more buttons depending on which page you
are on.

There will be a button to take you back to the welcome page or, on the
welcome page itself, refresh the newsgroups list to show the current
count.

When logged in as a moderator there will be a button to take you to, or
to refresh, the article list.

When logged in as admin there will be a button to take you back to the
management page if you are not there already.

The mechanism has been set up to be a step towards allowing moderators
to also have admin rights so a user with both moderation and management
rights would get buttons for both.

Also includes some tests.