tests for processWebRequest and move tests to t directory
This adds a test of the processWebRequest function covering all the
incoming actions for the different users or lack of user. This function
is the primary place where authentication and authorization is done so
tests are needed before any work can be done in this area.

Tests for perl are conventionally kept under a directory named 't' so
the existing tests have been moved there along with the new test. This
should help if there is a move towards using the normal perl build/test
tools.