Management of user rights
This change allows users to be granted admin rights. The rights have
been separated into two - listAdmin and userAdmin -  with a right -
fullAdmin - that includes both. The structure is intended to allow
further refinement or addition of actions wit other kinds of right. The
'ADMIN' user has been retained with fixed 'fullAdmin' rights.

The management page for a userAdmin/fullAdmin shows user rights and
allows updates. User data is saved to a 'users' file rather than
'moderators' with automatic upgrade of an existing moderators file.

Change password is now by button in the table of users.

Add user is by a row at the bottom of the table of users which allows
rights to be set for the new user.

Delete user is by a button in the row corresponding to a user - the
'ADMIN' user may not be deleted.

A user with 'listAdmin' rights but not 'userAdmin' does not see the
table of users on the management page, just the forms related to the
various moderation actions. Similarly, a user with 'userAdmin' but not
'listAdmin' rights does not see the forms related to moderation actions.

The new code has been put into modules with corresponding tests.

User admin forms moved to new module and updated to provide an add user
line and delete buttons for the existing users (except ADMIN).

Separated user admin and list admin and reflect that on the management
page so the user sees only the forms they can use.

More tests, including testing the rights separation yet.