Hide quoted lines in multi article display
On the moderation page that shows multiple articles, lines that start
'>' as is standard for quoting are replaced by a count of how many such
lines have been hidden and neither the lines nor the count are included
for the purposes of the 5 line limit.

The count and the hidden lines are wrapped in DIV elements which should
make it relatively straightforward to add buttons and scripts to
collapse and expand the hidden lines.